backend/requirements.txt
Django>=3.2,<4.0
psycopg2-binary
geopandas
rasterio
gdal~=3.10.3
djangorestframework
djangorestframework-gis
celery
redis
django-environ
owslib
lxml
bmi-topography
soilgrids
requests
